requirements.txt
alembic==1.5.8
astroid==2.5.3
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.0.4
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
protobuf==3.17.3
PyJWT==2.0.1
pylint==2.7.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.19.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.7
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
